third_party/blink/web_tests/external/wpt/css/css-align/blocks/align-content-block-002-expected.txt

This is a testharness.js-based test.
[FAIL] .test 5: last baseline
  assert_equals: \n<div class="test" style="align-content: last baseline" title="last baseline">\n    <div class="in-flow" data-offset-y="41"></div>\n    <div class="in-flow">\n      <span class="label">LAST BASELINE</span>\n      <span class="abspos" data-offset-y="26">ABS</span>\n      <span class="relpos" data-offset-y="26">REL</span>\n      <div class="overflow">OVERFLOW</div>\n    </div>\n  </div>\noffsetTop expected 41 but got 10
Harness: the test ran to completion.
